Move functions out of core.js
