+++ src/headless/plugins/roster/presence.js
@@ -2,7 +2,7 @@ import isNaN from "lodash-es/isNaN";
 import { Collection } from "@converse/skeletor/src/collection";
 import { Model } from '@converse/skeletor/src/model.js';
 import { converse } from "@converse/headless/core";
-import { initStorage } from '@converse/headless/shared/utils.js';
+import { initStorage } from '@converse/headless/utils/storage.js';
 
 const { Strophe, dayjs, sizzle } = converse.env;
