+++ src/headless/plugins/roster/utils.js
@@ -1,7 +1,7 @@
 import log from "@converse/headless/log";
 import { Model } from '@converse/skeletor/src/model.js';
 import { _converse, api, converse } from "@converse/headless/core";
-import { initStorage } from '@converse/headless/shared/utils.js';
+import { initStorage } from '@converse/headless/utils/storage.js';
 
 const { $pres } = converse.env;
