+++ src/headless/utils/storage.js
@@ -1,8 +1,5 @@
-import DOMPurify from 'dompurify';
 import Storage from '@converse/skeletor/src/storage.js';
-import log from '@converse/headless/log';
 import { _converse, api } from '@converse/headless/core';
-import { getOpenPromise } from '@converse/openpromise';
 
 export function getDefaultStore () {
     if (_converse.config.get('trusted')) {
@@ -35,29 +32,4 @@ export function initStorage (model, id, type) {
         model.on('destroy', () => window.removeEventListener(_converse.unloadevent, flush));
         model.listenTo(_converse, 'beforeLogout', flush);
     }
-}
-
-export function replacePromise (name) {
-    const existing_promise = _converse.promises[name];
-    if (!existing_promise) {
-        throw new Error(`Tried to replace non-existing promise: ${name}`);
-    }
-    if (existing_promise.replace) {
-        const promise = getOpenPromise();
-        promise.replace = existing_promise.replace;
-        _converse.promises[name] = promise;
-    } else {
-        log.debug(`Not replacing promise "${name}"`);
-    }
-}
-
-const element = document.createElement('div');
-
-export function decodeHTMLEntities (str) {
-    if (str && typeof str === 'string') {
-        element.innerHTML = DOMPurify.sanitize(str);
-        str = element.textContent;
-        element.textContent = '';
-    }
-    return str;
 }
