+++ src/shared/constants.js
@@ -1,0 +1,27 @@
+// These are all the view-layer plugins.
+//
+// For the full Converse build, this list serves
+// as a whitelist (see src/converse.js) in addition to the
+// CORE_PLUGINS list in src/headless/consts.js.
+export const VIEW_PLUGINS = [
+    'converse-bookmark-views',
+    'converse-chatboxviews',
+    'converse-chatview',
+    'converse-controlbox',
+    'converse-dragresize',
+    'converse-fullscreen',
+    'converse-headlines-view',
+    'converse-mam-views',
+    'converse-minimize',
+    'converse-modal',
+    'converse-muc-views',
+    'converse-notification',
+    'converse-omemo',
+    'converse-profile',
+    'converse-push',
+    'converse-register',
+    'converse-roomslist',
+    'converse-rootview',
+    'converse-rosterview',
+    'converse-singleton'
+];
